+++ apps/docs/src/app/platform/component-docs/platform-forms/switch/switch-header/switch-header.component.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'fdp-switch-header',
+    templateUrl: './switch-header.component.html'
+})
+export class SwitchHeaderComponent {}