+++ libs/core/src/lib/combobox/combobox.component.ts
@@ -317,7 +317,6 @@ export class ComboboxComponent implements ComboboxInterface, ControlValueAccesso
             if (this.searchFn) {
                 this.searchFn();
             }
-            this._moveCursorToInputEnd();
         } else if (KeyUtil.isKey(event, 'ArrowDown')) {
             if (event.altKey) {
                 this._resetDisplayedValues();
@@ -411,7 +410,7 @@ export class ComboboxComponent implements ComboboxInterface, ControlValueAccesso
     }
 
     /** @hidden */
-    onPrimaryButtonClick(event: MouseEvent): void {
+    onPrimaryButtonClick(): void {
         // Prevent primary button click behaviour on mobiles
         if (this.mobile) {
             return;
@@ -420,8 +419,6 @@ export class ComboboxComponent implements ComboboxInterface, ControlValueAccesso
         if (this.searchFn) {
             this.searchFn();
         }
-        event.preventDefault();
-        event.stopPropagation();
         this._resetDisplayedValues();
         this.isOpenChangeHandle(!this.open);
         this.searchInputElement.nativeElement.focus();
@@ -458,8 +455,10 @@ export class ComboboxComponent implements ComboboxInterface, ControlValueAccesso
 
     /** Method that handles complete event from auto complete directive, setting the new value, and closing popover */
     handleAutoComplete(event: AutoCompleteEvent): void {
-        this.inputText = event.term;
-        this.handleSearchTermChange();
+        if (this.inputText !== event.term) {
+            this.inputText = event.term;
+            this.handleSearchTermChange();
+        }
         if (event.forceClose) {
             this.isOpenChangeHandle(false);
         }
@@ -473,7 +472,6 @@ export class ComboboxComponent implements ComboboxInterface, ControlValueAccesso
                 forceClose: false
             });
         }
-        this._moveCursorToInputEnd();
     }
 
     /** Method that picks other value moved from current one by offset, called only when combobox is closed */
@@ -570,8 +568,8 @@ export class ComboboxComponent implements ComboboxInterface, ControlValueAccesso
         try {
             this.focusTrap = focusTrap(this._elementRef.nativeElement, {
                 clickOutsideDeactivates: true,
-                returnFocusOnDeactivate: true,
-                escapeDeactivates: false
+                escapeDeactivates: false,
+                initialFocus: this._elementRef.nativeElement
             });
         } catch (e) {
             console.warn('Unsuccessful attempting to focus trap the Combobox.');
@@ -591,12 +589,6 @@ export class ComboboxComponent implements ComboboxInterface, ControlValueAccesso
     /** @hidden */
     private _hasDisplayedValues(): boolean {
         return this.open && this.displayedValues && this.displayedValues.length > 0;
-    }
-
-    /** @hidden */
-    private _moveCursorToInputEnd(): void {
-        const value = this.searchInputElement.nativeElement.value;
-        this.searchInputElement.nativeElement.setSelectionRange(value.length, value.length);
     }
 
     /** @hidden */