+++ libs/core/src/lib/form/form-control/form-control.directive.ts
@@ -40,15 +40,13 @@ export class FormControlDirective implements CssClassBuilder, OnInit, OnChanges 
      * function must return single string
      * function is responsible for order which css classes are applied
      */
-    buildComponentCssClass(): string {
+    buildComponentCssClass(): string[] {
         return [
             this.state ? 'is-' + this.state : '',
             this._getFormClass(),
             this.compact ? this._getFormClass() + '--compact' : '',
             this.class
-        ]
-            .filter((x) => x !== '')
-            .join(' ');
+        ];
     }
 
     private _getFormClass(): string {