+++ libs/core/src/lib/multi-input/multi-input.component.ts
@@ -275,14 +275,12 @@ export class MultiInputComponent implements
      * function must return single string
      * function is responsible for order which css classes are applied
      */
-    buildComponentCssClass(): string {
+    buildComponentCssClass(): string[] {
         return [
             'fd-multi-input',
             'fd-multi-input-custom',
             this.class
-        ]
-            .filter((x) => x !== '')
-            .join(' ');
+        ];
     }
 
     elementRef(): ElementRef<any> {
@@ -320,6 +318,9 @@ export class MultiInputComponent implements
 
     /** @hidden */
     openChangeHandle(open: boolean): void {
+        if (this.disabled) {
+            return ;
+        }
         if (this.open !== open) {
             this.openChange.emit(open);
         }
@@ -541,6 +542,4 @@ export class MultiInputComponent implements
     private shouldFilterValues(changes): boolean {
         return this.dropdownValues && (changes.dropdownValues || changes.searchTerm);
     }
-
-
 }